Change activity names and add shim registry
