--- packages/firebase_ml_vision/example/android/app/src/main/java/io/flutter/plugins/firebasemlvisionexample/MainActivity2.java
@@ -1,12 +1,0 @@
-package io.flutter.plugins.firebasemlvisionexample;
-
-import io.flutter.embedding.android.FlutterActivity;
-import io.flutter.embedding.engine.FlutterEngine;
-import io.flutter.plugins.firebasemlvision.FirebaseMlVisionPlugin;
-
-public class MainActivity2 extends FlutterActivity {
-  @Override
-  public void configureFlutterEngine(FlutterEngine flutterEngine) {
-    flutterEngine.getPlugins().add(new FirebaseMlVisionPlugin());
-  }
-}
